Make sure that decoded value is an array
It is technically valid to encode a scalar value in JSON. In that case, there are no decoding errors (so you get `JSON_ERROR_NONE` when decoding), but as the data is not an array, the call to `$request->request->replace()`` will result in an internal server error, as `replace()` is typehinted to expect an array as argument 1. Therefore, an explicit check is required and added with this commit.